requirements.txt
web3>=6.0.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.0.0
click>=8.0.0
